fix: 분석 정합성 검증 로직 강화 및 영속성 동기화 버그 수정
- hashing 로직을 utils/hashUtils로 중앙 집중화하여 모든 트리거에서 동일한 해시를 사용하도록 보증
- pendingDocuments를 IndexedDB에 영속화하여 사이트 재접속 시에도 분석 중인 문서의 상태를 유지
- EditorPage 저장 시 분석 버퍼에 자동으로 업데이트를 전달하여 자동 분석 정합성 확보
- WorldPage 분석 버튼에 중복 방지 및 버퍼 체크 로직 추가
